fix: move axis drag scaling into YAxis/XAxis widgets
